Some bug fixes & formatting corrections for NPB test scripts.

+++ bench/NPB3.3-OMP-C/scripts/external_workload.sh
@@ -10,7 +10,7 @@ NUM_ITERATIONS=50
 BENCH_SIZE="medium"
 OMP_NUMA_AWARE_MAPPING=0
 EXTERNAL_WORKLOADS="1 2 4 8 16"
-BREATHER=0
+BREATHER=5
 
 ###############################################################################
 ## Helper functions
@@ -21,11 +21,11 @@ function print_help {
 	echo
 	echo "Usage: external_workload.sh [ OPTIONS ]"
 	echo "Options:"
-	echo -e "\t-h/--help : print help & exit"
+	echo -e "\t-h/--help   : print help & exit"
 	echo -e "\t-i <number> : number of iterations to run (default is $NUM_ITERATIONS)"
 	echo -e "\t-s <size>   : benchmark size (valid values are S, W, A, B, C, medium & all)"
 	echo -e "\t              default is $BENCH_SIZE"
-	echo -e "\t-n          : enable NUMA-aware mapping decisions* (default is $OMP_NUMA_AWARE_MAPPING"
+	echo -e "\t-n          : enable NUMA-aware mapping decisions* (default is $OMP_NUMA_AWARE_MAPPING)"
 	echo
 	echo "*Not yet implemented!"
 	exit 0
@@ -37,6 +37,9 @@ function print_help {
 
 function run_workload_launcher {
 	local bench_seq="$1"
+
+	# Cut out DC, as it doesn't play nicely with other benchmarks
+	bench_seq=${bench_seq/dc.W.x}
 
 	# Generate long list
 	for i in `seq 5`; do
@@ -60,6 +63,9 @@ function run_external_workload {
 		*) error "Unknown bench size $BENCH_SIZE" ;;
 	esac
 
+	# Cut out LU, as it takes an astronomically long time to co-run
+	benches=${benches/lu.A.x}
+
 	# Iterate through different external workload cases
 	for external_workload in $EXTERNAL_WORKLOADS; do
 		for bench in $benches; do
@@ -68,20 +74,24 @@ function run_external_workload {
 			fi
 
 			# Start background launchers
-			for i in `seq 2 $external_workload`; do
-				run_workload_launcher "$benches" &
-			done
-			sleep $BREATHER # Wait for the external workload to get moving
+			if [ $external_workload -ne 1 ]; then
+				for i in `seq 2 $external_workload`; do
+					run_workload_launcher "$benches" &
+				done
+				sleep $BREATHER # Wait for the external workload to get moving
+			fi
 
 			# Run current bench
-			for i in `seq $NUM_ITERATIONS`; do
-				run_bench $bench $i $RESULTS/${bench}-${external_workload}-${iteration}.log
+			for iter in `seq $NUM_ITERATIONS`; do
+				run_bench $bench $i $RESULTS/${bench}-${external_workload}-${iter}.log
 			done
 
 			# Kill external workload
-			pkill -TERM -P $$
-			sleep $BREATHER # Wait for external workload to finish
-			shared_mem_clear
+			if [ $external_workload -ne 1 ]; then
+				pkill -TERM -P $$
+				sleep $BREATHER # Wait for external workload to finish
+			fi
+			shared_mem_clear # Just in case...
 		done
 	done
 }
